✨ Add user service to API
