+++ server/src/user/user.module.ts
@@ -1,4 +1,8 @@
 import { Module } from '@nestjs/common';
+import { UserService } from './services/user.service';
 
-@Module({})
+@Module({
+  providers: [UserService],
+  exports: [UserService],
+})
 export class UserModule {}
